feature(api): Move geography model methods in repository
